requirements.txt
aniso8601==9.0.1
attrs==22.2.0
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
coverage==7.2.1
exceptiongroup==1.1.1
flasgger==0.9.5
Flask==2.2.3
Flask-Caching==2.0.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
packaging==23.0
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytz==2022.7.1
PyYAML==6.0
rfc3339-validator==0.1.4
six==1.16.0
SQLAlchemy==2.0.6
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
